ml_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Data & ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
lightgbm==4.3.0
xgboost==2.0.3
imbalanced-learn==0.12.2
shap==0.45.0
optuna==3.6.1
networkx==3.3
joblib==1.4.2
scipy==1.13.0

# Visualisation (training scripts only)
matplotlib==3.8.4
seaborn==0.13.2
